search fix

+++ BusssinessObject/ProductBusiness.cs
@@ -23,7 +23,7 @@ namespace BusssinessObject
         Task<IBusinessResult> DeleteById(int id);
         Task<bool> IdExists(int id);
         IQueryable<List<Product>> GetProductsAsQueryable();
-        Task<IBusinessResult> SearchProduct(string? query, decimal? minPrice, decimal? maxPrice, List<int>? categoryIDs, string? condition, bool? isAvailable, long? sellerID, int pageIndex = 1, int pageSize = 10);
+        Task<IBusinessResult> SearchProduct(string? query, decimal? minPrice, decimal? maxPrice, List<int>? categoryIDs, string? condition, bool? isAvailable, int? sellerID, int pageIndex = 1, int pageSize = 10);
 
     }
     public class ProductBusiness : IProductBusiness
@@ -46,7 +46,7 @@ namespace BusssinessObject
         {
             return _unitOfWork.ProductRepository.GetProductsAsQueryable();
         }
-        public async Task<IBusinessResult> SearchProduct(string? query, decimal? minPrice, decimal? maxPrice, List<int>? categoryIDs, string? condition, bool? isAvailable, long? sellerID, int pageIndex = 1, int pageSize = 10)
+        public async Task<IBusinessResult> SearchProduct(string? query, decimal? minPrice, decimal? maxPrice, List<int>? categoryIDs, string? condition, bool? isAvailable, int? sellerID, int pageIndex = 1, int pageSize = 10)
         {
             try
             {
